Add script to automate BoringSSL rebuilds.
"Knows" the correct arguments for each supported
architecture and can automate `git pull` etc.

Updated BUILDING.md to document it.

Also updated the Windows build instructions
as there were for a deprecated version of MSVC.

Also fixed a very minor bug in the openjdk
build.gradle which would cause if to crash
on empty PATH elements (more common on Windows).